forum examen (2)/forum examen/backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
flask-socketio==5.6.0
pandas==3.0.0
openpyxl==3.1.5
bcrypt==5.0.0
mysql-connector-python==9.6.0
python-socketio==5.16.1
